refactor: Moved common files
