+++ src/drive/web/modules/views/SelectionBarWithActions.jsx
